Extract ItemEditorView into Separate File
